Fix test expectations when libvirtd is off

+++ libvirt/tests/src/virsh_cmd/domain/virsh_undefine.py
@@ -260,7 +260,11 @@ def run(test, params, env):
     # Check results.
     if status_error:
         if not status:
-            test.fail("virsh undefine return unexpected result.")
+            if libvirtd_state == "off" and libvirt_version.version_compare(5, 6, 0):
+                logging.info("From libvirt version 5.6.0 libvirtd is restarted "
+                             "and command should succeed")
+            else:
+                test.fail("virsh undefine return unexpected result.")
         if params.get('setup_libvirt_polkit') == 'yes':
             if status3 == 0:
                 test.fail("virsh define with false acl permission" +
